Add faster Binary protocol implementation to read/write structs
Summary:
This is similar to Compact extensions which provide methods to directly read/write structs using string. These methods are faster than the existing ones due to decreased calls to Hack for interacting with string buffer in TMemoryBuffer.

While Binary protocol is not used as heavily as the Compact protocol, this may help with one-off use cases that use Binary in streaming like MySQL client. Until we figure out a long term fix for serialization extensions, this should help with the perf.

Differential Revision: D87884195

fbshipit-source-id: 1a6f826027043bf1454cdd8e5f1374efa1765179