Treat HH\is_any_array as the same as is HH\AnyArray<_,_>
Summary: We've eliminated `is_vec($x)` refinement in favour of `$x is vec<_>` (and similar for `dict` and `keyset`) but `is_any_array` remains. Let's just treat it the same as `is HH\AnyArray<_,_>` for the purposes of refinement. (Existing treatment produces bad arity uses of `not AnyArray` in the `else` branch, for unknown reasons.)

Reviewed By: viratyosin

Differential Revision: D87919429

fbshipit-source-id: e471bc7c26095b6ec62799e541c761ae82e410ee