Add carbon::result::OVERLOADED
Summary:
Similar to `carbon::result::HOT_KEY`, we want to keep the same logic, except we don't want to fail over.

This is a generic result and should be used in case of overload protection. This result does not fail over, and is treated as a "hard" error. This result will be sent by UCache when it is in an overloaded state to protect itself.

Reviewed By: stuclar

Differential Revision: D87820744

fbshipit-source-id: 04c4b3c430e6fbfa3f7e89f1813d463c3a72eadb