vsdebug: Fix double-closing listener FDs
While working on D88186184, I noticed that we were closing the listener FDs
for the vsdebug server in both `waitForConnection` and `listenForClientConnection`.
It does not seem to be causing issues in local testing but it could be a
problem if the FD gets reused in a different thread in the meantime.
`waitForConnection` is only called at the end of
`listenForClientConnection` so closing the FDs only at the end of
`listenForClientConnection` should be fine.