Fix `StructuredCursorReader::readField` dangling pointer bug.
Summary:
As titled.

Before this fix, this method had a bug when `U` in the input `U& val` was a `std::string_view` (some terse fields also be affected). In this case, copying the underlying data converted it into a temporary `std::string` that would immediately by cleaned up after binding it to a `std::string_view`. This left the `std::string_view` with a dangling reference to the temporary string we created.

This change fixes it by deferring the copy of copyable types until assignment/return.

Reviewed By: iahs

Differential Revision: D88119409

fbshipit-source-id: 93e6c62c43ccaf3005cd6389634b8665eea453ce