Add a test case for function arguments with optional arguments
Summary:
This diff adds a test case for a function that takes a function with an optional parameter.

When calling the function with common mistakes (e.g. missing the optional parameter, or passing optional parameter as a mandatory). The error messages are technically correct but most likely will confuse the user.

There are two issues with the snippet:
- Too many arguments are emitted twice (with 4041 and 4105).
- Error message could be more helpful and point to the issue with optional arguments.

Reviewed By: andrewjkennedy

Differential Revision: D88312788

fbshipit-source-id: b4c9988c845acc0e35b20dc1ade5a54b16fe809c