Support type constant super bounds in polymorphic function pointerss
Summary:
Adds support for type constant super bounds when synthesizing types for polymorphic function pointers.

The main subtlety with this code come from nested type constant access e.g.

```
abstract class One {
  abstract const type Ta as Two super Three;
  public function foo(this::Ta::Tb): void {}
}

abstract class Two {
  abstract const type Tb as ... super ...;
}

abstract class Three extends Two {
  abstract const type Tb as ... super ...;
}

```

To cope with the scenario where upper and lower bounds both define a nested constant this diff extends the existing code in a couple of ways:

1) Since the typeconstant can now have both an upper and lower bound we need to ensure they refined to agress on any nested type constant access in the function signature.

This is done by taking the intersection of the upper bounds and the union of the lower bounds.

Concrete definitions are treated as the the point interval.

2) The process above leads to intersection and union types so we extend refinement of type parameters generated when moving existentials to universal position to support them.

For union types, we require that all elements of the union define the type constant. Each receives the same refinement to ensure they agree.

For intersection types, we require that at least one element of the intersection defines the type constant. Only elements which define the type constant will have a refinement applied.

In general, a user could declare upper and lower bounds with nested type constants with disjoint bounds. In keeping with the approach in the rest of Hack this doesn't cause an error - it just means the user has written dead code.

Reviewed By: andrewjkennedy

Differential Revision: D88069095

fbshipit-source-id: 31a886c19f1c7c0c27b3c31dbeef636db4ae927d