Add an __ImplementedBy attribute for methods on primitives
Summary:
For methods on primitives, we want to give a class in an .hhi, and tie that to the top-level function that is the runtime implementation of the method. We also want to ensure that the types are compatible. We add an attribute (only allowed in .hhi files) that declares the name of the top-level function. We then fund  the function's decl and check that it is a subtype of the method's type. The methods type is instantiated as for `meth_caller` to take into account generics from the class. For example, in the .hhi for string, we might have:
```
class string {
  <<__ImplementedBy('\my_length')>>
  public method length(): int;
}

function my_length(string $s): int;
```

We use the `Fun_id.synth` and `Meth_caller.synth` functions from `typing.ml` to build the types of the functions. In particular, the Meth_caller one adds the "this" parameter for us, and moves class generics to function generics. It also tries to add `readonly` and `FunctionRef` wrappings, which we disable for this purpose.

Reviewed By: mheiber

Differential Revision: D88174803

fbshipit-source-id: 5308745bf57a91e71b94fa4eeaac54b7adc64fa1