Internally, permit splat to follow optional in tuples
Summary:
Function parameters can be marked optional (with defaults for functions and concrete methods, or using the `optional` keyword for abstract methods), and tuple elements can be marked optional. In both cases, we impose the restriction that optional parameters may not precede required parameters, as Hack does not support the notion of "missing arguments" or "missing elemens" e.g. one could imagine a syntax such as `foo(2,,4)` or `tuple(2,,4)`.

This design led to a further restriction for type-splats in functions and tuples: an optional parameter or element may not precede a type-splat, because instantiating a type-splat with at least one required parameter or element would break the invariant described above. For example, if a function has the signature `foo<T as (mixed...)>(int $x, optional string $y, ...T $rest):void` then instantiating `T` with `(int,string)` would produce a function type in which an optional parameter preceded required parameters.

Unfortunately this rules out some useful functionality, brought to light with typed format strings, where many functions have an optional format string followed currently by a variadic but in future a type-splat argument.

Therefore we propose relaxing this restriction, at least internally, permitting function types and tuple types to have a type-splat element following optional parameters or elements. When the type-splat is "instantiated", optional parameters that precede required parameters will be "interpreted" as required, i.e. the "gap is filled in".

This diff just changes the representation of tuple types: instead of a list of required parameters followed by *either* optional parameters + variadic *or* a type-splat, the tuple type is a list of required parameters, and then optional parameters, and then a variadic *or* a type-splat. i.e. it changes from

`required; ((optional; variadic) | splat)`

to

`required; optional; (variadic | splat)`

where `;` denotes sequencing and `|` denotes alternation.

There is no new functionality yet i.e. the existing restrictions apply. Some of the functionality (e.g. subtyping) needs extending to be expressive enough to support relaxation of the restrictions.

Reviewed By: enetsee

Differential Revision: D88077436

fbshipit-source-id: cd0d0e80b2df328082b559e29f347a9d87a57295