Improve __ImplementedBy checking
Summary: Instead of hacking Meth_caller.synth to not wrap the type in FunctionRef, split out a separate top-level function. Also call the top-level function synthesis that wraps readonly and FunctionRef for the implementation side. This is harmless because `FunctionRef<T> <: T`. But because function ref isn't covariant, we cannot have it on the RHS. Using the top-level synthesis like this will hopefully let us support having a real function pointer in the attribute instead of a string in the future.

Reviewed By: andrewjkennedy

Differential Revision: D88415517

fbshipit-source-id: 2fc89a7ed9a073f81700a1256a2a3e2cf3233095