increase size of stacks used by rust_decl_ffi thread pool
Summary:
This diff addresses a problem where we overflow the stack when running `Concurrent::step` in `rust_decl_ffi.rs`.
In there, we run the direct decl parser in a `rayon` thread pool. These threads have a default stack size of 2MB, which is too small for the parser.

This diff makes sure that we do not use the implicitly configured global `rayon` thread pool, but create our own one that is initialized upon first use and shared by all instances of `Concurrent`. I'm using the `stack_size` option from the parser option. Its default is 32MB, which is enough for us (see test plan). Interestingly, it seems like this setting was previously unused.

Currently, the code in `direct_decl_ffi.rs` is only used by the `ClientIdeDaemon` process.

Alternatives considered:
- In theory, we could use arbitrarily large values for the stack size, since the memory is only reserved as needed (i.e., as each individual stack grows). But that would make the reported `VSZ` of our IDE processes very large.
- Instead of creating our own thread pool, we could configure the global pool to have larger stacks. However, this would mean that we have to make sure that if any new code paths may reach `Concurrent::step` in the future, they also have to do this configuration correctly, and there is no way to check if this happened.
- `Concurrent` instances are explicitly created by calling `hh_concurrent_parse_start_ffi` from OCaml first. This means that we could attach one thread pool to each `Concurrent` instance, rather than having one shared by all instances. However, the `start` method seems to be called frequently from `Direct_decl_prefetch.prefetch_decls` without old instances being cleaned up first, which means that we may end of up with a lot of thread pools in parallel. So let's just have a single one.

Reviewed By: nt591

Differential Revision: D88477177

fbshipit-source-id: a714006be3e1ea54289ae17532065f92e1b6e3fb