Create a __SoftRequirePackage attribute
Summary:
This attribute is sort of the marriage between __RequirePackage and __Deprecated. Rather than asserting the package exists, we check if it exists and if it doesn't we trigger a log in the same way as __Deprecated, including sampling, 0-for-no-logs, etc.

I really don't like the
`class_name: Option<StringId>` and `self.class_name = Some(intern(class.name.name()));` stuff, but I couldn't come up with another way to do this without removing one or more of Copy, Clone, and Default from the struct traits. Strings don't play nicely with copy/clone, which is lovely.

Reviewed By: vassilmladenov

Differential Revision: D88268835

fbshipit-source-id: 1321b48d7f74da77dbd29ca4c9cec926615b3dd1