Make string final
Summary: We want string to be final. Since it isn't a real class at runtime, it makes no sense to try to extend it. This required an exception to be inserted in the typing union code where string shouldn't be treated as minimal even though it is final, because things like `classname<C>` are non-trivial subtypes of string.

Reviewed By: viratyosin

Differential Revision: D88481599

fbshipit-source-id: d17da5131171b7157fada91f2ed2c57524b4daf9