StressTest: Create a new runner framework for rpc comparison
Summary:
Need a new runner framework for comparing Thrift with Grpc (and potentially other rpcs in the future).
We are not using the existing conformance setup because the stress test client's implementation here differs from what that setup uses, due to the way grpc works.
Some refactoring might be possible, but it's going to be a much bigger change due to the cross-cell dependencies, so for now it seemed cleaner to just have a different setup for this.

Reviewed By: robertroeser

Differential Revision: D88443898

fbshipit-source-id: c1ce3796f84f0ce57b35106695d62e697266c64a