Restrict __ImplementedBy to final classes
Summary: The logic in meth_caller for building a function type for a method isn't in general sound when dealing with the `this` type. However, it is sound for `final` classes. `__ImplementedBy` relies on this logic, so we restrict it to final classes, which are the only classes that we need (or will support) for methods on primitives

Reviewed By: andrewjkennedy

Differential Revision: D88484290

fbshipit-source-id: 4a12ea3cb5944d4eaa87ebd677110566eb661cce