Remove useless and misleading warning in PHPOutputTransport destructor
Summary:
PHPOutputTransport is used in 4 well controlled places in the extension that
trivially call flush() or onewayFlush() at the end.

The only possible situation where we fail to flush the buffer is if
PHPOutputTransport gets destructed due to an Exception. In which case not
flushing it is intentional and we should not emit a misleading warning.

Reviewed By: ricklavoie

Differential Revision: D88527731

fbshipit-source-id: 74919870d32ee8ad5dcf7d94a2a02bb2ba60e440