support for enums
Summary:
This diff adds support for enums to `hh --find-my-tests`. This includes ordinary enums and enum classes.

The selection heuristics are pretty simple: If an enum `MyEnum` is given as a root (i.e., was changed by a diff), we look for any method that contains any kind of reference to the type `MyEnum`, and continue our usual traversal from there. This means that there is no special treatment for `MyEnum::getNames` and friends.

Implementation:
- Within `ServerFindMyTests`, our existing notion of `Classish` nodes now carries around a `Ast_defs.classish_kind` to distinguish our previously supported "normal" classes from enums and enum clases.
- On the command line, we now allow passing classes to `--find-my-tests` for the first time, using similar syntax `Class|MyEnum` as used for `--find-refs`.
- Also inspired by `--find-refs`, all that parsing is now handled in `ClientCheck`, we give an `action` value to `ServerFindMyTests` instead of doing any parsing there.

Reviewed By: madgen

Differential Revision: D88103118

fbshipit-source-id: 65031a1551f391c792f9292658855d69987ab522