Template ContainerTraits and DefaultValueWriter on ProtocolWriter
Summary:
Template ContainerTraits::write methods and DefaultValueWriter on ProtocolWriter
instead of hardcoding BinaryProtocolWriter. This is part of incremental work to make
cursor-based serialization support multiple protocols. ContainerTraits::write is now a
template method, and DefaultValueWriter is templated on both Tag and ProtocolWriter with
BinaryProtocolWriter as the default. Updated all usages to include the ProtocolWriter
template parameter.

Reviewed By: praihan

Differential Revision: D88508189

fbshipit-source-id: ec11c0056020e487297a77894a112b29ced9095a