Add headers to TxnBytesEvent for LAST_HEADER_BYTE_READ
Summary: Includes the HTTP headers in the `TxnBytesEvent` when triggering the `LAST_HEADER_BYTE_READ` event. This allows `HTTPTransactionObserver` implementations to access and inspect response headers (or request headers, depending on context) at the moment the last header byte is read, enabling more detailed observation and logging of transaction metadata.

Reviewed By: aharonnovo

Differential Revision: D88845675

fbshipit-source-id: f4004a80c31939465a07e5ab1e5cffa65ffee0b0