psp: Make KernelPSP interface asynchronous
Summary:
Change the KernelPSP interface to return folly::SemiFutures and update the
AsyncPSPUpgrade state machine to handle the asynchronous completion of
PSP operations.

The kernel PSP implementation may perform sleepable (blocking) operations
during the userspace netlink send(). This can happen, for instance, if the
device is slow to respond and/or the command queue corresponding to the
device is full.

Since AsyncPSPUpgrade runs in an I/O context, this blocking behavior is
undesirable.

Differential Revision: D88776382

fbshipit-source-id: c36d68622d731ae38f2b7709b19d6626a6570413