Utility to validate schema of thrift blob
Summary:
Thrift's core model is that both writer and reader know the schema of the messages they send, so asking whether a blob corresponds to a given schema is generally ill-formed. However, there is utility to this when developing systems that implement custom dynamic Thrift serialization, such as TXL. These systems can use such a utility to validate that the outputs they produce are compatible with the schema they used, allowing some bugs in those systems to be caught.
This diff adds a utility that determines whether the value could have been produced by the schema, differentiating backward-compatible evolution (adding fields) from incompatible evolution (two fields with the same identity but different types).
The return value is chosen to emphasize the fact that this should not be used to positively identify the types of blobs, such as to store untagged unions of Thrift types in databases.

Reviewed By: sadroeck, praihan

Differential Revision: D88693896

fbshipit-source-id: 898fd1935236cba51ec124ce09ea424c31741d1e