Make more of exception_shared_string constexpr
Summary:
Up-stack, rich error code has fairly robust `constexpr` support. The major benefit for me is that constant-evaluated code is run by the compiler in a VM that actively blocks UB. So, the portions of the code that are exercised at compile-time end up demonstrably safe, even more so than UBSAN can give us.

It's also nice, of course, to have a reasonable error-handling paradigm for `constexpr` code before C++26.

Reviewed By: ispeters

Differential Revision: D88685982

fbshipit-source-id: 76b24372a05d809fb3ae2cc78d0da52e4b0fa2ae