amortize acquire order in hazptr_domain::load_hazptr_vals
Summary:
The function loads all of the values from an array of atomics each with memory-order-acquire, and then uses all of them.

This is equivalent to loading all of the values each with memory-order-relaxed, following with a fence with memory-order-acquire, and then using all of the values. But on aarch64, where loads by-default are relaxed and load-acquire costs an extra few cycles each time, we can use a single fence to amortize the acquire ordering across all of the loads.

With this change, there are no more per-hprec load-acquire operations when traversing the hprec's. There are only a fixed number of load-acquire and fence-acquire operations where the acquire orderings amortize across all of the loads.

Reviewed By: paulmckrcu

Differential Revision: D84995209

fbshipit-source-id: 4efb564602e886baaddd9c5dd66cab213f939b9e