use StaticSingletonManager for default_hazptr_domain
Summary:
May fix some crashes we've seen where the `ObserverManager` refresh thread segfaults in `hazptr_domain::push_list` in `hazptr_obj_base::retire`. This call is the first moment where the domain storage is accessed.

The hypothesis is that the `default_hazptr_domain` is in a DSO and the `ObserverManager` is in either another DSO or the main executable. In this situation, the DSO containing `default_hazptr_domain` is unloaded and its storage unmapped, so that the next access to its storage segfaults. This happens while the main executable or the second DSO is still alive and the `ObserverManager` refresh thread is still running.

This is a slightly hacky workaround since only the storage for the default domain sticks around after the DSO containing it is unloaded, not any of the functions in that DSO. But all of the hazptr-related functions are inline, so maybe it will be okay.

Regarding performance, this should be okay. The measurable existing cost of calling `default_hazptr_domain()` is very low. It's presumably low because it lowers to object code that compares an address with a relocated immediate. So actually loading a bit from an offset in the domain will be more costly, but probably not by much in an absolute sense.

Reviewed By: ot

Differential Revision: D88852802

fbshipit-source-id: b71c70d49200375c68d17bd7649efa4af826ec6f