optimize hazptr_holder::try_protect
Summary:
On aarch64, a store-release operation lowers to `stlr` while a store-relaxed operation lowers to `str`, and `stlr` costs a bit more than `str`.

When lightweight fences are available, it is possible to optimize a store-release into the sequence fence-light-release then store-relaxed, as long as what the combo synchronizes-with pessimizes its load-acquire into the sequence load-relaxed then fence-heavy-acquire.

In this case, `hazptr_rec::reset_hazptr()` is extremely hot in the hazptr readers, and just performs the store-release. It synchronizes-with collection of the hazard pointers in `hazptr_domain::load_hazptr_vals()`, which is extremely cold. It may prove a worthwhile tradeoff to optimize `hazptr_rec::reset_hazptr()` and to pay for the optimization with a second heavyweight fence in the hazptr reclaimer.

This effect is measurable in the hazptr micro-benchmarks. It remains to be seen whether this is visible at scale, in production scenarios.

Reviewed By: paulmckrcu

Differential Revision: D85697666

fbshipit-source-id: 158e81bd7701bd5f1c9fda34b8a3add83cb1a73c