Mark reified class-level generics as erased during function pointer elaboration
Summary:
This diff removes a restriction on the use of `meth_caller` on instance methods of classes with reified type parameters. The crux of the change is as follows; given the following class, with reified type parameter `T`:

```
class C<reify T> {
   public function __construct(private T $t): void {}

   public function foo(T $t): void { ... }
}
```

The expression `meth_caller(C::class, 'foo')` is currently given the following function type:

```
function<reify T>(C<T>, T): void
```

and the change means it would instead be given the following type  *without the `reify` specifier`* :

```
function<T>(C<T>, T): void
```

The current typing means we reject the `meth_caller` expression or require a FIXME which we think is unnecessary, hence the new typing.

andrewjkennedy gave a nice intuition motivating this by looking at `reify` in terms of an elaboration.

Where we have a *function / method type parameter marked as `reify`* we can look at it as elaborating to a function / method with an additional parameter of type `TypeStructure<...>`, for example, given:

```
class C {
  public function bar<reify T>(T $t): void { ... }
}
```

we can think of `bar` as follows:

```
  public function bar<T>(T $t, TypeStructure<T>) $ts_t): void { ... }
```

Where we have a *class type parameter marked as `reify`* the elaboration actually happens on the constructor, reusing the previous example:

```
class C<refiy T> {
    public function __construct(private T $t): void {}
}
```

can be looked at as:

```
   public function __construct(private T $t,public TypeStructure $ts_t): void {}
```

The point here is that when we an instantiate the class `C` we _must_ have provided `TypeStructure<T>` and that information is available when we call instance methods.

Reviewed By: andrewjkennedy

Differential Revision: D89280578

fbshipit-source-id: 20324a539148a5db501588bfbbdbdf04c1bee8f8