Add a config option to control release of methods on strings
Summary: Add a `length` method to `string`, and top-level `__primitive_length` to `string.hhi`. The `ignore_string_method` config option (default true) causes the direct decl parser and type checker to ignore any methods on string and functions in string.hhi. This will let us experiment with the feature and control the roll-out.

Reviewed By: enetsee

Differential Revision: D88380459

fbshipit-source-id: 11ff80954ede9a1c1da407ce90cb60999ae3f0a5