Fix O(N^2) blow up with long AddElemC chains during DCE in HHBBC
Summary:
As part of the final phase of HHBBC, we perform DCE on the
now optimized unit. This DCE relies on liveness information, which is
inferred by walking the program *backwards*. Since bytecode ops are,
in general, not reversible, this creates a problem. This is solved
with the "undo log", which keeps a log of all pushes/pops/mutations
when running the abstract interpreter forwards. Once the final state
is arrived at, the undo log can then by replayed to compute the
resulting states going backwards.

This all works fairly well, but long AddElemC chains can be
problematic. In such a long chain, we'll have a single array which we
repeatedly add a key and value to. We go to great effort to prevent
the array from having more than one ref-count, so that each AddElemC
mutates the array type in place. Otherwise, each mutation will trigger
COW and we end up doing O(N^2) work in length of the array
chain. (Each mutation will need to copy the current chain length). For
large arrays this quickly becomes prohibitive.

However, in order to undo such AddElemCs, the undo log stores the
before AddElemC array (repeatedly), thereby ensuring that every
mutation triggers a COW and we get the O(N^2) slowdown.

Fix this by adding special casing to the undo log. When the array is
an ArrLikeMap specialization, and the key is known statically, we
don't actually need to store the array in the undo log. In this
situation, we sufficies to just store the key and some other
information. Then, we can take the mutated array (which is already
present on the stack), and undo the mutation by removingg the value
with the remembered key. ArrLikeMap is the only specialization where
this matters. The rest don't grow in proportion to the AddElemC
chain. Static arrays (where everything about the array is known) are
already dealt with by a different mechanism.

This speeds up final-pass by a few minutes because we're spending the
majority of time doing DCE on these big AddElemC chains.

Reviewed By: sinancepel

Differential Revision: D89108553

fbshipit-source-id: 0705db051a2eedaf120e14499eeb0c9aa29cb48c