remove re-entrancy when serializing objects from Xenon
Summary:
The xenon extension added an output file option and would call
json_encode to reach into the variable serializer which in turn calls
jsonSerialie. This was re-entrant and observed with a debugger breakpoint
stepping into serialization methods.

This diff tweaks variable-serializer to take a flag to not call jsonSerialize
on the object, instead using the normal object serialization. We then use that
code in Xenon.

Reviewed By: paulbiss

Differential Revision: D88662741

fbshipit-source-id: b960643509836ec55539f44487971103d08e68c2