Fix projection off tuple with optional elements
Summary: Generalize tuple projection in subtyping to deal with case of optional tuple elements preceding a splat element, which will shortly be enabled. (We will add some tests.) At the same time, fix a small bug whereby we don't reject a coalescing index that always "fails". (We reject the corresponding operation on shapes.)

Differential Revision: D89273443

fbshipit-source-id: c3966f80a622151a9b7f326fc375226e5cff8cd9