While Hovering/Pressing Support
Summary:
### Summary

This diff enhances the ProtoKitUI library by introducing support for hovering and pressing interactions. The changes include:

Introduction of a new script, `PKUIFlowUnhoveringHandler.cs`, which implements the `IHoverHandler` interface to manage hover interactions. PKUIFlowUnhoveringHandler is created on Start by the FrameInteraction that has a trigger == Pressing || Hovering and is parented to the target frame

There are two main way to "go back from a Pressing/Hovering" transition

if the Transition In is finished we siply call Navigate back like normal with a opposite transition,

if the trasnition is still running I've a  introduced in `PKUIFrameTransition.cs` a way to Interrupt and rewind. There's work work that could be done to clean this up further

Note : Overlay interruption needs to be fixed in a further diff

Differential Revision: D86207508

fbshipit-source-id: a34fcbccff9a89c7a341af0cfe41992e1bbfab88