Some improvements to HHBBC parallel library
Summary:
If you're calling parallel functions in an inner loop, the overhead of
creating and destroying the threads can easily dominate the
runtime. For this use case, add a thread_pool class, which allows you
to pre-create the threads and reuse them inside the loop. If a
thread_pool isn't provided, the functions will create their own
threads, as before.

If the current thread is already inside a parallel call, prevent
nested calls to parallel. This is done with a thread local global,
which is set in the worker threads. If you call any parallel functions
while within a parallel, the inner call will execute the work
serially. There's a couple of places in HHBBC which (unintentionally)
call parallel functions while inside parallel threads. This can cause
an explosion of threads and is now prevented.

Reviewed By: sinancepel

Differential Revision: D89327833

fbshipit-source-id: 838fe00f3512a697a77346260df91563c8f20aec