Improve blob encoding of std::unique_ptr and copy_ptr
Summary:
Initially I was wary of automatically serde-ing pointer types, as
there was some question about the proper behavior. Instead, you needed
to opt-in using a macro to emit the right serialization on a type by
type basis. This causes problems in a few places and I'm now more
confident that trying to encode the inner type is always right.

So, drop the macros and have first-class support for std::unique_ptr
and copy_ptr.

I also noticed that serde-ing a SHA1 took up more space than
expected. That's because the integer zig-zag encoding is actually
larger than the raw bits (because the bits have high entropy). So, for
SHA1 encode the hash directly, not as a set of integers.

Reviewed By: sinancepel

Differential Revision: D89354087

fbshipit-source-id: 9069730063bfb895060c73c314f8aceb87f95532