benchmark hazptr reclamation with many retired objects
Summary:
In particular, take care to ensure that the retired objects are evicted from cache before reclamation, so that we can benchmark reclamation in the expected cold-cache scenario. Reclamation almost never traverses retired objects repeatedly across runs, with the exception of objects which have very long-lived protection critical sections, and this can be simulated with a cold cache.

Note that the existing reclamation benchmarks intend to measure traversal of the hazard pointers, which can be expected to be in cache when reclamation runs. In the steady state, reclamation traverses the same set of hazptr_rec's in the same order each time.

Reviewed By: skrueger

Differential Revision: D88042809

fbshipit-source-id: ee265a9b102a446c88316fba290bc8755b803cb2