Reduce extreme CPU usage during heavy translate lease contention
Summary:
If there is a giant function that results in many profiling or live
translations and there are many concurrent requests trying to execute it, we
would keep spending most CPU time on getContext() and RegionContext destructor,
since all attempts will fail to obtain the lease.

Seems that this issue was accidentally introduced by D48879852 that hoisted the
expensive getContext() call out of the lease.

This needs to be fixed further for async jitting and defer getContext() call
if we already enqueued the item.

Reviewed By: EdwaRen, mdko

Differential Revision: D89485960

fbshipit-source-id: 4228455d021d4ab604bcacc94ce8d6dd8d9d234c