let hazptr_domain::do_reclamation interleave sharded linked-list walks
Summary:
Linked list iteration intrinsically has a loop-carried dependency. When the lists are long, the dependency chain can get long and fill the processor's execution window, stalling execution and costing cycles.

It turns out that the retired-pointer linked lists (`hazptr_domain::untagged_` and `hazptr_domain::tagged_`) are already sharded. The primary purpose of sharding these lists is to accelerate retirement by relieving contention on the unsharded lists. But we can take advantage of the sharding in reclamation as well.

Reclamation current iterates these lists one after the other. This does nothing to relieve the dependency chains that form when iterating each list. But if reclamation were to interleave iteration the linked list shards, then the dependency chains could run in parallel. That would hopefully not fill the processor's execution window and not stall execution.

Reviewed By: skrueger

Differential Revision: D87689685

fbshipit-source-id: 1ad22c68f7d57df2d8ad33d15570ba7033a8346b