use folly::available_concurrency instead of sysconf(_SC_NPROCESSORS_ONLN)
Summary:
This change replaces calls of sysconf(_SC_NPROCESSORS_ONLN) with
folly::available_concurrency when the caller wants the number of processors
available to a job.

This should be largely a no-change since most jobs use platform010 with glibc
2.34 where these functions behave identically.  However, jobs on Aarch64 use
glibc 2.40.  There, the folly method continues to return the number of
processors available to a job but sysconf(_SC_NPROCESSORS_ONLN) returns the
number of processors known to the operating system, a different number when the
job is stacked.

Most of fbcode expects the folly behavior and this codemod only
affects those callers.  By making the switch to the folly method, code
will behave the same when code is run on Aarch64 or when the glibc is
upgraded on x86.

Reviewed By: yfeldblum

Differential Revision: D88771423

fbshipit-source-id: f921eb0b3f59a4333e55d9890f544ce85bb916b8