Unset unnamed locals after ?as expressions
Summary:
For nullable `as` expressions on failure we were neglecting to unset the `type_struct_local` and `arg_local` on the failure path when pushing `null`. We don't actually need `type_struct_local` in the nullable case so this diff avoids assigning it and in both the enforcing and nullable cases the `arg_local` can be replaced with a `Dup` instruction.

Separately, in a few scenarios it's possible for us to throw while preforming an `IsTypeStructC` instruction which would leave the local initialized with the type structure dangling in the event of a catch.

Differential Revision: D89770958

fbshipit-source-id: 57580d125bb60f6885bf636a5c5714d0c51aac48