Fix detection of oldest JIT worker thread when retranslate-all is stuck
Summary:
There was logic to kill (and get a stacktrace for) the oldest JIT worker threads
when the main retranslate-all thread get stuck, but that would only kill the
oldest JIT worker if it too had exceeded the stuck-detection limit.  This
diff changes the logic to kill the oldest JIT worker regardless of that when
the oldest thread is the main retransalate-all thread.

Reviewed By: mdko

Differential Revision: D89823982

fbshipit-source-id: 70cd78a24e956672b66c4ff24f14ad34284e78fb